Fix codex warnings.

+++ libpandabase/utils/dfx.cpp
@@ -72,7 +72,7 @@ void DfxController::ResetOptionValueFromString(const std::string &s)
         last_pos = s.find_first_not_of(';', pos);
         pos = s.find(';', last_pos);
         std::string option_str = arg.substr(0, arg.find(':'));
-        uint8_t value = std::stoi(arg.substr(arg.find(':') + 1));
+        uint8_t value = static_cast<uint8_t>(std::stoi(arg.substr(arg.find(':') + 1)));
         auto dfx_option = DfxOptionHandler::DfxOptionFromString(option_str);
         if (dfx_option != DfxOptionHandler::END_FLAG) {
             DfxController::SetOptionValue(dfx_option, value);
